refactor(pkg): separated types from Pilot.ts
This keeps things cleaner and creates more division between what's internal or not. uwu
