+++ packages/pilot/src/regex-router.ts
@@ -1,8 +1,7 @@
 /**
  * © 2022 WavePlay <dev@waveplay.com>
  */
-import { PilotRouter, PilotRouteResult, PilotRouterOptions } from './pilot';
-import { PilotRouteOptions } from './route';
+import type { PilotRouter, PilotRouteOptions, PilotRouteResult, PilotRouterOptions } from './types';
 import { match } from 'node-match-path';
 
 export class RegexRouter implements PilotRouter {
